refactor(publicacion.controller): make query construction more flexible
The query construction in the `getPublicacionesByTag` function is now more flexible by conditionally adding the `tag` field only if it exists. This improves maintainability and prevents potential issues when the `tag` query parameter is not provided. The error message is also generalized to be more inclusive of different query scenarios.